created separate function for saveTrail validator
